fix after SETSe update. Peels mean tension per node
This update fixes all the embedding code after the major update to the way that SETse functions.
I also added in a change to the way the the Peels quintet is projected. Tension is no mean per node. this is a much better way of reflecting the relationship of nodes with thier neighbours